Extract config interfaces into separate file
