+++ src/ts/demofactory.ts
@@ -24,7 +24,7 @@ import {VolumeToggleButton} from './components/volumetogglebutton';
 import {PlaybackToggleButton} from './components/playbacktogglebutton';
 import {SeekBar} from './components/seekbar';
 import {VideoQualitySelectBox} from './components/videoqualityselectbox';
-import {UIConditionContext, UIConfig, UIManager} from './uimanager';
+import {UIConditionContext, UIManager} from './uimanager';
 import {TitleBar} from './components/titlebar';
 import {BufferingOverlay} from './components/bufferingoverlay';
 import PlayerAPI = bitmovin.PlayerAPI;
@@ -33,6 +33,7 @@ import {AudioTrackListBox} from './components/audiotracklistbox';
 import {SettingsPanelItem} from './components/settingspanelitem';
 import {SettingsPanelPage} from './components/settingspanelpage';
 import { UIFactory } from './uifactory';
+import { UIConfig } from './uiconfig';
 
 export namespace DemoFactory {
